+++ sim_tools/__init__.py
@@ -1,5 +1,5 @@
 """sim-tools"""
-__version__ = '0.9.0'
-__author__ = 'Thomas Monks'
+__version__ = '0.10.0'
+# __author__ removed - now in pyproject.toml
 
 from . import datasets, distributions, time_dependent, ovs, output_analysis