+++ tests/test_dists_unit.py
@@ -241,15 +241,40 @@ def test_truncated_min(n, lower_bound):
     """
     Check that samples from the TruncatedDistribution do not fall below the
     specified lower bound.
-
-    Parameters
-    ----------
-    n: int
-        Number of samples to generate from the distribution.
-    lower_bound: float
-        The lower bound value that should be enforced by the truncated
-        distribution.
     """
     d1 = dists.Normal(10, 1, random_seed=SEED_1)
     d2 = dists.TruncatedDistribution(d1, lower_bound=lower_bound)
     assert min(d2.sample(size=n)) >= lower_bound
+
+
+def test_registry_batch_sorting():
+    """Check that DistributionRegistry.create_batch() sorting works."""
+    d_config = {
+        "b_dist": {"class_name": "Exponential", "params": {"mean": 1}},
+        "a_dist": {"class_name": "Exponential", "params": {"mean": 1}}
+    }
+    d_sorted = dists.DistributionRegistry.create_batch(d_config, sort=True)
+    d_unsorted = dists.DistributionRegistry.create_batch(d_config, sort=False)
+    assert list(d_sorted.keys()) == ["a_dist", "b_dist"]
+    assert list(d_unsorted.keys()) == ["b_dist", "a_dist"]
+
+
+@pytest.mark.parametrize("conf, should_pass", [
+    ({"class_name": "Exponential", "params": {"mean": 1}}, True),
+    ({"class_name": "Exponential"}, False),
+    ({"params": {"mean": 1}}, False),
+    ({"class_name": "Exponential", "params": {"mean": 1}, "foo": 123}, False),
+    ({"CLASS_NAME": "Exponential", "params": {"mean": 1}}, False),
+])
+def test_registry_batch_validation(conf, should_pass):
+    """
+    Check that DistributionRegistry.create_batch() warns for unsuitable
+    distribution configurations.
+    """
+    seed = 123
+    if should_pass:
+        obj = dists.DistributionRegistry._validate_and_create(conf, seed)
+        assert hasattr(obj, "sample")
+    else:
+        with pytest.raises(ValueError):
+            dists.DistributionRegistry._validate_and_create(conf, seed)